feat(apps): put settings link and page behind auth
